Add files via upload
This code is for testing one moisture sensor only, along with logging data on an SD card. It takes reading from one moisture sensor, displays into on the serial monitor and writes it onto a file in the SD card. Code was successfully tested. Could be useful to test single moisture sensors, or SD cards, and can be built on for more sensors.